commit main branch comments 2/16/21>>>>//:
The first version of FREE YanSan is very close to being compete and this latest commit has restructured the main loop drastically.  I have built a Main class for all the group updates, blits, drawings, and 'if' statements in order to clean the mess I had.  This has improved game speed a lot!